Refactor successfully BarcodeAnalyzer in its own class
